+++ src/EventEmitter.ts
@@ -1,13 +1,13 @@
-import { NativeModules, NativeEventEmitter } from "react-native";
-import type { EventSubscription } from "react-native";
-import type { AdEventObject, AdEventListener } from "./types/AdEvent";
+import { NativeModules, NativeEventEmitter } from 'react-native';
+import type { EventSubscription } from 'react-native';
+import type { AdEventObject, AdEventListener } from './types/AdEvent';
 
 const { AppLovinMAX } = NativeModules;
 
 // Note that this is a singleton in ES6 module
 const emitter = new NativeEventEmitter(AppLovinMAX);
 
-const subscriptions: Record<string, EventSubscription> = { };
+const subscriptions: Record<string, EventSubscription> = {};
 
 export const addEventListener = <T extends AdEventObject>(event: string, handler: AdEventListener<T>) => {
     const subscription: EventSubscription = emitter.addListener(event, handler);
@@ -16,7 +16,7 @@ export const addEventListener = <T extends AdEventObject>(event: string, handler
         currentSubscription.remove();
     }
     subscriptions[event] = subscription;
-}
+};
 
 export const removeEventListener = (event: string) => {
     const currentSubscription = subscriptions[event];
@@ -24,4 +24,4 @@ export const removeEventListener = (event: string) => {
         currentSubscription.remove();
         delete subscriptions[event];
     }
-}
+};