+++ src/InterstitialAd.ts
@@ -1,8 +1,8 @@
-import { NativeModules } from "react-native";
-import { addEventListener, removeEventListener } from "./EventEmitter"
-import type { AdDisplayFailedInfo, AdInfo, AdLoadFailedInfo, AdRevenueInfo } from "./types/AdInfo";
-import type { LocalExtraParameterValue } from "./types/AdProps";
-import type { InterstitialAdType } from "./types/InterstitialAd";
+import { NativeModules } from 'react-native';
+import { addEventListener, removeEventListener } from './EventEmitter';
+import type { AdDisplayFailedInfo, AdInfo, AdLoadFailedInfo, AdRevenueInfo } from './types/AdInfo';
+import type { LocalExtraParameterValue } from './types/AdProps';
+import type { InterstitialAdType } from './types/InterstitialAd';
 
 const { AppLovinMAX } = NativeModules;
 
@@ -18,83 +18,81 @@ const {
 
 const isAdReady = (adUnitId: string): Promise<boolean> => {
     return AppLovinMAX.isInterstitialReady(adUnitId);
-}
+};
 
 const loadAd = (adUnitId: string): void => {
     AppLovinMAX.loadInterstitial(adUnitId);
-}
+};
 
-const showAd = (
-    adUnitId: string,
-    placement?: string | null,
-    customData?: string | null
-): void => {
+const showAd = (adUnitId: string, placement?: string | null, customData?: string | null): void => {
     AppLovinMAX.showInterstitial(adUnitId, placement ?? null, customData ?? null);
-}
+};
 
 const setExtraParameter = (adUnitId: string, key: string, value: string | null): void => {
     AppLovinMAX.setInterstitialExtraParameter(adUnitId, key, value);
-}
+};
 
 const setLocalExtraParameter = (adUnitId: string, key: string, value: LocalExtraParameterValue): void => {
     AppLovinMAX.setInterstitialLocalExtraParameter(adUnitId, { [key]: value });
-}
+};
 
 const addAdLoadedEventListener = (listener: (adInfo: AdInfo) => void) => {
     addEventListener(ON_INTERSTITIAL_LOADED_EVENT, (adInfo: AdInfo) => listener(adInfo));
-}
+};
 
 const removeAdLoadedEventListener = () => {
     removeEventListener(ON_INTERSTITIAL_LOADED_EVENT);
-}
+};
 
 const addAdLoadFailedEventListener = (listener: (errorInfo: AdLoadFailedInfo) => void) => {
     addEventListener(ON_INTERSTITIAL_LOAD_FAILED_EVENT, (errorInfo: AdLoadFailedInfo) => listener(errorInfo));
-}
+};
 
 const removeAdLoadFailedEventListener = () => {
     removeEventListener(ON_INTERSTITIAL_LOAD_FAILED_EVENT);
-}
+};
 
 const addAdClickedEventListener = (listener: (adInfo: AdInfo) => void) => {
     addEventListener(ON_INTERSTITIAL_CLICKED_EVENT, (adInfo: AdInfo) => listener(adInfo));
-}
+};
 
 const removeAdClickedEventListener = () => {
     removeEventListener(ON_INTERSTITIAL_CLICKED_EVENT);
-}
+};
 
 const addAdDisplayedEventListener = (listener: (adInfo: AdInfo) => void) => {
     addEventListener(ON_INTERSTITIAL_DISPLAYED_EVENT, (adInfo: AdInfo) => listener(adInfo));
-}
+};
 
 const removeAdDisplayedEventListener = () => {
     removeEventListener(ON_INTERSTITIAL_DISPLAYED_EVENT);
-}
+};
 
 const addAdFailedToDisplayEventListener = (listener: (errorInfo: AdDisplayFailedInfo) => void) => {
-    addEventListener(ON_INTERSTITIAL_AD_FAILED_TO_DISPLAY_EVENT, (errorInfo: AdDisplayFailedInfo) => listener(errorInfo));
-}
+    addEventListener(ON_INTERSTITIAL_AD_FAILED_TO_DISPLAY_EVENT, (errorInfo: AdDisplayFailedInfo) =>
+        listener(errorInfo)
+    );
+};
 
 const removeAdFailedToDisplayEventListener = () => {
     removeEventListener(ON_INTERSTITIAL_AD_FAILED_TO_DISPLAY_EVENT);
-}
+};
 
 const addAdHiddenEventListener = (listener: (adInfo: AdInfo) => void) => {
     addEventListener(ON_INTERSTITIAL_HIDDEN_EVENT, (adInfo: AdInfo) => listener(adInfo));
-}
+};
 
 const removeAdHiddenEventListener = () => {
     removeEventListener(ON_INTERSTITIAL_HIDDEN_EVENT);
-}
+};
 
 const addAdRevenuePaidListener = (listener: (adInfo: AdRevenueInfo) => void) => {
     addEventListener(ON_INTERSTITIAL_AD_REVENUE_PAID, (adInfo: AdRevenueInfo) => listener(adInfo));
-}
+};
 
 const removeAdRevenuePaidListener = () => {
     removeEventListener(ON_INTERSTITIAL_AD_REVENUE_PAID);
-}
+};
 
 export const InterstitialAd: InterstitialAdType = {
     isAdReady,
@@ -124,6 +122,6 @@ export const InterstitialAd: InterstitialAdType = {
 
     addAdRevenuePaidListener,
     removeAdRevenuePaidListener,
-}
+};
 
 export default InterstitialAd;