+++ src/index.ts
@@ -1,13 +1,13 @@
-export { default, AppLovinMAX } from "./AppLovinMAX";
-export { Privacy } from "./Privacy";
-export { TargetingData, AdContentRating, UserGender } from "./TargetingData";
-export { InterstitialAd } from "./InterstitialAd";
-export { RewardedAd } from "./RewardedAd";
-export { AppOpenAd } from "./AppOpenAd";
-export { BannerAd } from "./BannerAd";
-export { MRecAd } from "./MRecAd";
-export { AdView, AdFormat, AdViewPosition } from "./AdView";
-export { NativeAdView } from "./nativeAd/NativeAdView";
+export { default, AppLovinMAX } from './AppLovinMAX';
+export { Privacy } from './Privacy';
+export { TargetingData, AdContentRating, UserGender } from './TargetingData';
+export { InterstitialAd } from './InterstitialAd';
+export { RewardedAd } from './RewardedAd';
+export { AppOpenAd } from './AppOpenAd';
+export { BannerAd } from './BannerAd';
+export { MRecAd } from './MRecAd';
+export { AdView, AdFormat, AdViewPosition } from './AdView';
+export { NativeAdView } from './nativeAd/NativeAdView';
 export {
     TitleView,
     AdvertiserView,
@@ -17,5 +17,5 @@ export {
     OptionsView,
     MediaView,
     StarRatingView,
-} from "./nativeAd/NativeAdViewComponents";
-export * from "./types";
+} from './nativeAd/NativeAdViewComponents';
+export * from './types';