+++ src/nativeAd/NativeAdView.tsx
@@ -1,23 +1,25 @@
-import React, { forwardRef, useContext, useImperativeHandle, useRef, useState, useEffect, useCallback } from "react";
-import { NativeModules, requireNativeComponent, UIManager, findNodeHandle } from "react-native";
-import type { ViewProps } from "react-native";
-import { NativeAdViewContext, NativeAdViewProvider } from "./NativeAdViewProvider";
-import type { AdInfo, AdLoadFailedInfo, AdRevenueInfo } from "../types/AdInfo";
-import type { AdNativeEvent } from "../types/AdEvent";
-import type { NativeAd } from "../types/NativeAd";
-import type { NativeAdViewHandler, NativeAdViewProps } from "../types/NativeAdViewProps";
-import type { NativeAdViewType, NativeAdViewContextType } from "./NativeAdViewProvider";
+import React, { forwardRef, useContext, useImperativeHandle, useRef, useState, useEffect, useCallback } from 'react';
+import { NativeModules, requireNativeComponent, UIManager, findNodeHandle } from 'react-native';
+import type { ViewProps } from 'react-native';
+import { NativeAdViewContext, NativeAdViewProvider } from './NativeAdViewProvider';
+import type { AdInfo, AdLoadFailedInfo, AdRevenueInfo } from '../types/AdInfo';
+import type { AdNativeEvent } from '../types/AdEvent';
+import type { NativeAd } from '../types/NativeAd';
+import type { NativeAdViewHandler, NativeAdViewProps } from '../types/NativeAdViewProps';
+import type { NativeAdViewType, NativeAdViewContextType } from './NativeAdViewProvider';
 
 const { AppLovinMAX } = NativeModules;
 
 type NativeAdViewNativeEvents = {
-    onAdLoadedEvent(event: { nativeEvent: { nativeAd: NativeAd; adInfo: AdInfo; } }): void
-    onAdLoadFailedEvent(event: AdNativeEvent<AdLoadFailedInfo>): void
-    onAdClickedEvent(event: AdNativeEvent<AdInfo>): void
-    onAdRevenuePaidEvent(event: AdNativeEvent<AdRevenueInfo>): void
+    onAdLoadedEvent(event: { nativeEvent: { nativeAd: NativeAd; adInfo: AdInfo } }): void;
+    onAdLoadFailedEvent(event: AdNativeEvent<AdLoadFailedInfo>): void;
+    onAdClickedEvent(event: AdNativeEvent<AdInfo>): void;
+    onAdRevenuePaidEvent(event: AdNativeEvent<AdRevenueInfo>): void;
 };
 
-const NativeAdViewComponent = requireNativeComponent<NativeAdViewProps & ViewProps & NativeAdViewNativeEvents>('AppLovinMAXNativeAdView');
+const NativeAdViewComponent = requireNativeComponent<NativeAdViewProps & ViewProps & NativeAdViewNativeEvents>(
+    'AppLovinMAXNativeAdView'
+);
 
 /**
  * The {@link NativeAdView} component that you use building a native ad. This loads a native ad and
@@ -34,7 +36,7 @@ const NativeAdViewComponent = requireNativeComponent<NativeAdViewProps & ViewPro
  * {@link NativeAdView} fills each asset view with the data of a native ad as soon as it loads the native
  * ad, but you need to provide the layout and style of the asset views.
  * {@link NativeAdView} can reload a new native ad by using the ref handler.
- * 
+ *
  * ### Example:
  * ```js
  * <NativeAdView
@@ -56,46 +58,52 @@ const NativeAdViewComponent = requireNativeComponent<NativeAdViewProps & ViewPro
  * </NativeAdView>
  * ```
  */
-export const NativeAdView = forwardRef<NativeAdViewHandler, NativeAdViewProps & ViewProps>(function NativeAdView(props, ref) {
-    const [isInitialized, setIsInitialized] = useState<boolean>(false);
+export const NativeAdView = forwardRef<NativeAdViewHandler, NativeAdViewProps & ViewProps>(
+    function NativeAdView(props, ref) {
+        const [isInitialized, setIsInitialized] = useState<boolean>(false);
 
-    useEffect(() => {
-        // check that AppLovinMAX has been initialized
-        AppLovinMAX.isInitialized().then((result: boolean) => {
-            setIsInitialized(result);
-            if (!result) {
-                console.warn("ERROR: NativeAdView is mounted before the initialization of the AppLovin MAX React Native module");
-            }
-        });
-    }, []);
+        useEffect(() => {
+            // check that AppLovinMAX has been initialized
+            AppLovinMAX.isInitialized().then((result: boolean) => {
+                setIsInitialized(result);
+                if (!result) {
+                    console.warn(
+                        'NativeAdView is mounted before the initialization of the AppLovin MAX React Native module'
+                    );
+                }
+            });
+        }, []);
 
-    // Not ready to render NativeAdView
-    if (!isInitialized) {
-        return null;
+        // Not ready to render NativeAdView
+        if (!isInitialized) {
+            return null;
+        }
+
+        return (
+            <NativeAdViewProvider>
+                <NativeAdViewImpl {...props} ref={ref} />
+            </NativeAdViewProvider>
+        );
     }
+);
 
-    return (
-        <NativeAdViewProvider>
-            <NativeAdViewImpl {...props} ref={ref} />
-        </NativeAdViewProvider>
-    );
-});
-
-const NativeAdViewImpl = forwardRef<NativeAdViewHandler, NativeAdViewProps & ViewProps>(function NativeAdViewImpl({
-    adUnitId,
-    placement,
-    customData,
-    extraParameters,
-    localExtraParameters,
-    onAdLoaded,
-    onAdLoadFailed,
-    onAdClicked,
-    onAdRevenuePaid,
-    children,
-    style,
-    ...otherProps
-}, ref) {
-
+const NativeAdViewImpl = forwardRef<NativeAdViewHandler, NativeAdViewProps & ViewProps>(function NativeAdViewImpl(
+    {
+        adUnitId,
+        placement,
+        customData,
+        extraParameters,
+        localExtraParameters,
+        onAdLoaded,
+        onAdLoadFailed,
+        onAdClicked,
+        onAdRevenuePaid,
+        children,
+        style,
+        ...otherProps
+    },
+    ref
+) {
     // context from NativeAdViewProvider
     const { setNativeAd, setNativeAdView } = useContext(NativeAdViewContext) as NativeAdViewContextType;
 
@@ -107,7 +115,7 @@ const NativeAdViewImpl = forwardRef<NativeAdViewHandler, NativeAdViewProps & Vie
         if (nativeAdViewRef) {
             UIManager.dispatchViewManagerCommand(
                 findNodeHandle(nativeAdViewRef.current),
-                UIManager.getViewManagerConfig("AppLovinMAXNativeAdView").Commands.loadAd,
+                UIManager.getViewManagerConfig('AppLovinMAXNativeAdView').Commands.loadAd,
                 undefined
             );
         }
@@ -124,7 +132,7 @@ const NativeAdViewImpl = forwardRef<NativeAdViewHandler, NativeAdViewProps & Vie
         }
     }, []);
 
-    const onAdLoadedEvent = (event: { nativeEvent: { nativeAd: NativeAd; adInfo: AdInfo; } }) => {
+    const onAdLoadedEvent = (event: { nativeEvent: { nativeAd: NativeAd; adInfo: AdInfo } }) => {
         setNativeAd(event.nativeEvent.nativeAd);
         if (onAdLoaded) onAdLoaded(event.nativeEvent.adInfo);
     };