+++ src/types/BannerAd.ts
@@ -1,11 +1,10 @@
-import type { ViewAdType } from "./ViewAd";
-import type { AdViewPosition } from "../AdView";
+import type { ViewAdType } from './ViewAd';
+import type { AdViewPosition } from '../AdView';
 
 export type BannerAdType = ViewAdType & {
-
     /**
      * Creates a banner at the specified position and offsets.
-     * 
+     *
      * @param adUnitId The Ad Unit ID to load ads for.
      * @param position {@ AdViewPosition} position.
      * @param xOffset Horizontal offset from the left position.
@@ -14,8 +13,8 @@ export type BannerAdType = ViewAdType & {
     createAd(adUnitId: string, position: AdViewPosition, xOffset?: number, yOffset?: number): void;
 
     /**
-     * Sets a background color for the banner.  
-     * 
+     * Sets a background color for the banner.
+     *
      * @param adUnitId The Ad Unit ID to load ads for.
      * @param hexColorCode Hexadecimal color (#rrggbb).
      */
@@ -23,7 +22,7 @@ export type BannerAdType = ViewAdType & {
 
     /**
      * Sets the banner width.
-     * 
+     *
      * @param adUnitId The Ad Unit ID to load ads for.
      * @param width The desired banner width.
      */
@@ -31,7 +30,7 @@ export type BannerAdType = ViewAdType & {
 
     /**
      * Updates the banner position offsets.
-     * 
+     *
      * @param adUnitId The Ad Unit ID to load ads for.
      * @param xOffset Horizontal offset from the left position.
      * @param yOffset Vertical offset from the top position.
@@ -40,7 +39,7 @@ export type BannerAdType = ViewAdType & {
 
     /**
      * Gets the adaptive banner size for the provided width at the current orientation.
-     * 
+     *
      * @param width The banner width.
      */
     getAdaptiveHeightForWidth(width: number): Promise<number>;