+++ src/types/MRecAd.ts
@@ -1,11 +1,10 @@
-import type { ViewAdType } from "./ViewAd";
-import type { AdViewPosition } from "../AdView";
+import type { ViewAdType } from './ViewAd';
+import type { AdViewPosition } from '../AdView';
 
 export type MRecAdType = ViewAdType & {
-
     /**
      * Creates a MREC at the specified position.
-     * 
+     *
      * @param adUnitId The Ad Unit ID to  to load ads for.
      * @param position {@link AdViewPosition} position.
      */