+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Versions
 
 ## x.x.x
+* Update code format with `prettier`.
 * Update code with `eslint`.
 * Fix broken import path in `TargetingData`.
 * Fix wrong return type with `AppLovinMAX.setLocationCollectionEnabled()`.